+++ CHANGELOG.md
@@ -114,6 +114,7 @@ CHANGELOG
 - added `intelmq.bots.parsers.calidog.parser_certstream` for parsing certstream data (#1120).
 - added `intelmq.bots.parsers.shodan.parser` for parsing shodan data (#1096).
 - change the classification type from 'botnet drone' to infected system' in various parses.
+- `intelmq.bots.parsers.spamhaus.parser_cert`: Added support for all known bot types.
 
 #### Experts
 - Added sieve expert for filtering and modifying events (#1083)